I hate handling Client Data.
