+++ common/src/main/java/whocraft/tardis_refined/common/hum/TardisHums.java
@@ -1,7 +1,9 @@
 package whocraft.tardis_refined.common.hum;
 
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.sounds.SoundSource;
 import whocraft.tardis_refined.TardisRefined;
+import whocraft.tardis_refined.client.sounds.QuickSimpleSound;
 import whocraft.tardis_refined.common.util.CodecJsonReloadListener;
 import whocraft.tardis_refined.registry.TRSoundRegistry;
 
@@ -10,6 +12,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class TardisHums {
+
 
     private static final CodecJsonReloadListener<HumEntry> RELOAD_LISTENER = createReloadListener();
