+++ common/src/main/java/whocraft/tardis_refined/common/network/messages/sync/SyncHumsMessage.java
@@ -39,7 +39,7 @@ public class SyncHumsMessage extends MessageS2C {
 
     @Override
     public void toBytes(FriendlyByteBuf buf) {
-        buf.writeNbt((CompoundTag) (MAPPER.encodeStart(NbtOps.INSTANCE, this.tardisHums).result().orElse(new CompoundTag())));
+        buf.writeNbt(MAPPER.encodeStart(NbtOps.INSTANCE, this.tardisHums).result().orElse(new CompoundTag()));
     }
 
     @Override
